etl/pyproject.toml
[project]
name = "etl"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.5",
    "mysql-connector-python>=9.4.0",
    "pandas>=2.3.2",
    "psycopg-pool>=3.2.6",
    "psycopg2>=2.9.10",
    "psycopg[binary,c,pool]>=3.2.10",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "supabase>=2.21.1",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.13.1",
]
